+++ drift/test/delete_test.dart
@@ -71,23 +71,21 @@ void main() {
     });
   });
 
-  group('delete with from()', () {
+  group('delete on table instances', () {
     test('delete()', () async {
-      await db.from(db.users).delete().go();
+      await db.users.delete().go();
 
       verify(executor.runDelete('DELETE FROM users;', []));
     });
 
     test('deleteOne()', () async {
-      await db.from(db.users).deleteOne(const UsersCompanion(id: Value(3)));
+      await db.users.deleteOne(const UsersCompanion(id: Value(3)));
 
       verify(executor.runDelete('DELETE FROM users WHERE id = ?;', [3]));
     });
 
     test('deleteWhere', () async {
-      await db
-          .from(db.users)
-          .deleteWhere((tbl) => tbl.id.isSmallerThanValue(3));
+      await db.users.deleteWhere((tbl) => tbl.id.isSmallerThanValue(3));
 
       verify(executor.runDelete('DELETE FROM users WHERE id < ?;', [3]));
     });