+++ drift/test/insert_test.dart
@@ -454,4 +454,55 @@ void main() {
       ));
     });
   });
+
+  group('on table instances', () {
+    test('insert', () async {
+      await db.categories
+          .insert()
+          .insert(CategoriesCompanion.insert(description: 'description'));
+
+      verify(executor.runInsert(
+          'INSERT INTO categories ("desc") VALUES (?)', ['description']));
+    });
+
+    test('insertOne', () async {
+      await db.categories.insertOne(
+          CategoriesCompanion.insert(description: 'description'),
+          mode: InsertMode.insertOrReplace);
+
+      verify(executor.runInsert(
+          'INSERT OR REPLACE INTO categories ("desc") VALUES (?)',
+          ['description']));
+    });
+
+    test('insertOnConflictUpdate', () async {
+      when(executor.runSelect(any, any)).thenAnswer(
+        (_) => Future.value([
+          {
+            'id': 1,
+            'desc': 'description',
+            'description_in_upper_case': 'DESCRIPTION',
+            'priority': 1,
+          },
+        ]),
+      );
+
+      final row = await db.categories.insertReturning(
+          CategoriesCompanion.insert(description: 'description'));
+      expect(
+        row,
+        Category(
+          id: 1,
+          description: 'description',
+          descriptionInUpperCase: 'DESCRIPTION',
+          priority: CategoryPriority.medium,
+        ),
+      );
+
+      verify(executor.runSelect(
+        'INSERT INTO categories ("desc") VALUES (?) RETURNING *',
+        ['description'],
+      ));
+    });
+  });
 }