+++ drift/test/update_test.dart
@@ -180,4 +180,21 @@ void main() {
           ['new name', 3]));
     });
   });
+
+  group('update on table instances', () {
+    test('update()', () async {
+      await db.users.update().write(const UsersCompanion(id: Value(3)));
+
+      verify(executor.runUpdate('UPDATE users SET id = ?;', [3]));
+    });
+
+    test('replace', () async {
+      await db.categories.replace(const CategoriesCompanion(
+          id: Value(3), description: Value('new name')));
+
+      verify(executor.runUpdate(
+          'UPDATE categories SET "desc" = ?, priority = 0 WHERE id = ?;',
+          ['new name', 3]));
+    });
+  });
 }